mv prefix_of_string to kv_api
