+++ common/meta/types/proto/meta.proto
@@ -48,8 +48,10 @@ message WatchRequest {
   // key is the key to register for watching.
   string key = 1;
 
-  // key_end is the end of the range [key, key_end) to watch.
-  // if key_end is None, then watch only key.
+  // `key_end`` is the end of the range [key, key_end) to watch.
+  // If key_end is None, then watch only key.
+  // If want to watch prefix of key, use `get_start_and_end_of_prefix` to
+  // generate [key, key_end).
   optional string key_end = 2;
 
   enum FilterType {
